+++ mb_mgr.h
@@ -171,14 +171,6 @@ typedef struct MB_MGR {
         keyexp_t                keyexp_128;
         keyexp_t                keyexp_192;
         keyexp_t                keyexp_256;
-
-#ifndef NO_ADDON
-        /* state less cipher and hash add-on */
-        struct JOB_AES_HMAC * (*cipher_addon)(struct MB_MGR *, struct JOB_AES_HMAC *);
-        struct JOB_AES_HMAC * (*hash_addon)(struct MB_MGR *, struct JOB_AES_HMAC *);
-        void *user_data_0;
-        void *user_data_1;
-#endif /* !NO_ADDON */
 } MB_MGR;
 
 
@@ -318,18 +310,3 @@ get_next_job_sse(MB_MGR *state)
         (_mgr)->keyexp_192((_raw), (_enc), (_dec))
 #define IMB_AES_KEYEXP_256(_mgr, _raw, _enc, _dec) \
         (_mgr)->keyexp_256((_raw), (_enc), (_dec))
-
-#ifndef NO_ADDON
-static inline void
-set_mb_mgr_addon(MB_MGR *state,
-                 struct JOB_AES_HMAC * (*f_cipher_addon)(struct MB_MGR *, struct JOB_AES_HMAC *),
-                 struct JOB_AES_HMAC * (*f_hash_addon)(struct MB_MGR *, struct JOB_AES_HMAC *),
-                 void *user_data_0,
-                 void *user_data_1)
-{
-        state->cipher_addon = f_cipher_addon;
-        state->hash_addon = f_hash_addon;
-        state->user_data_0 = user_data_0;
-        state->user_data_1 = user_data_1;
-}
-#endif /* !NO_ADDON */
